[FIX] web: move test
The test added in dea66e1df317ca8782a68ece6c1e5917f6e8f3ec depends on a `mail`
helper, even though it is in the `web` module.

opw-3133731

X-original-commit: e0ac9c41c6852cc147a9dea236421193161d7abe
